refactor: move auth types to a type folder
