+++ src/api/application/authentication/types/index.ts
@@ -1,0 +1,9 @@
+export type AuthenticationRequest = {
+  username: string;
+  password: string;
+};
+
+export type AuthenticationResponse = {
+  accessToken: string;
+  expiresIn: number;
+};
